+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.1.0 - under development
 
+- Added Codeception test suite
 - Added `I18nActiveQuery` and extracted `ActiveRecord::$i18nAttributes` to `I18nAttributesTrait`
 
 ## 2.0.14 (Dec 8, 2023)